Function for factorial calculation is fixed. Permutation function is fixed too.

+++ common/combinatorics.h
@@ -1,6 +1,6 @@
 #ifndef COMBINATORICS_H_
 #define COMBINATORICS_H_
 
-unsigned long int fact(unsigned int N); 
+long int fact(long int N); 
 
 #endif
